03-aidd/requirements.txt
aiogram
openai
python-dotenv


